requirements.txt
Django==1.7.11
argparse==1.2.1
django-analytical==2.0.0
django-extensions==1.4.9
django-haystack==2.4.1
django-selectable==0.9.0
django-selenium==0.9.8
future==0.15.2
mongoengine==0.8.7
pefile==2016.3.28
psycopg2==2.5.4
pydot==1.0.2
pygraphviz==1.2
pymongo==2.7.2
pyparsing==1.5.7
python-ptrace==0.8.1
selenium==2.52.0
six==1.9.0
wsgiref==0.1.2
python-dateutil==2.4.2
pycountry==1.12
lxml==3.4.2
Unidecode==0.04.18
git+git://github.com/zeehio/django-kronos.git@f22f638120920b722edd7619d6a037289edc678e
git+git://github.com/jsteffen/django-jenkins.git@e84f76ee7c940f35f761ced65ed7c97aa00b4acb
httplib2==0.9.1
pygeoip==0.3.2
pysolr==2.1.0-beta
South==1.0.2
requests==2.7.0
